add:  リストのComponentを作成
